refactor(45): update component exports, routing and mocks
