+++ src/components/Header/Header.tsx
@@ -1,16 +1,25 @@
 import { Link } from 'react-router-dom';
 import { ROUTES } from '../../constants';
 
+const NAV_LINKS = [
+  { path: ROUTES.HOME, label: 'Home' },
+  { path: ROUTES.SHOP, label: 'Shop' },
+  { path: ROUTES.CART, label: 'Cart' },
+  { path: ROUTES.LOGIN, label: 'Login' },
+];
+
 export const Header = () => {
   return (
-    <nav>
-      <div>
-        <Link to={ROUTES.HOME}>Home</Link>
-        <Link to={ROUTES.SHOP}>Shop</Link>
-        <Link to={ROUTES.CONTACT_US}>Contact Us</Link>
-        <Link to={ROUTES.LOGIN}>Login</Link>
-        <Link to={ROUTES.CART}>Cart</Link>
-      </div>
-    </nav>
+    <header>
+      <nav>
+        <ul>
+          {NAV_LINKS.map(({ path, label }) => (
+            <li key={path}>
+              <Link to={path}>{label}</Link>
+            </li>
+          ))}
+        </ul>
+      </nav>
+    </header>
   );
 };
