+++ src/components/MainLayout/MainLayout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import { Header } from '../Header/Header';
+import { Header } from '../Header';
 
 export const MainLayout = () => {
   return (
