+++ src/components/index.ts
@@ -1,2 +1,3 @@
-export { Button } from './Button/Button';
-export { Header } from './Header/Header';
+export { Button } from './Button';
+export { Header } from './Header';
+export { MainLayout } from './MainLayout';
